requirements.txt
-i https://pypi.org/simple

alembic==1.5.4

dnspython==2.1.0; python_version >= '3.6'

email-validator==1.1.2

fastapi-sqlalchemy==0.2.1

fastapi==0.63.0

idna==3.1; python_version >= '3.4'

mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

psycopg2==2.8.6

psycopg2==2.8.6

pydantic[email]==1.7.3

python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-dotenv==0.15.0

python-editor==1.0.4

six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlalchemy==1.3.23; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

starlette==0.13.6; python_version >= '3.6'
PyJWT~=2.0.1
SQLAlchemy~=1.3.23
uvicorn~=0.13.3
bcrypt~=3.2.0
GeoAlchemy2~=0.8.4
geopy~=2.1.0
geojson~=2.5.0
Shapely~=1.7.1